test: fix e2e test access to settings with myWallet false
